* adjust dept field.
git-svn-id: file:///tmp/ranzhi/trunk@1001 cb094fd3-5a31-0410-bd9d-ef5a9cbf4261

+++ ranzhi/db/ranzhi.sql
@@ -441,7 +441,7 @@ CREATE TABLE `sys_task` (
 -- DROP TABLE IF EXISTS `sys_user`;
 CREATE TABLE `sys_user` (
   `id` smallint(5) unsigned NOT NULL AUTO_INCREMENT,
-  `dept` smallint(8) unsigned NOT NULL,
+  `dept` mediumint(8) unsigned NOT NULL,
   `account` char(30) NOT NULL DEFAULT '',
   `password` char(32) NOT NULL DEFAULT '',
   `realname` char(30) NOT NULL DEFAULT '',
